split batch code out of metrics ds
